+++ Runtime/DataStructs/Implementation/DataNodeBuilder/CombineRendererBuilder.cs
@@ -7,30 +7,38 @@ namespace DTTUnityCore.DataStructs
     public class CombineRendererBuilder : ICombineRendererBuilder
     {
         private MetaDataName _metaDataName;
-        private int _layer;
         private Dictionary<Material, List<CombineInstance>> _materialCombineInstanceDictionary;
+
+        private int _layer;
+        private bool _applyStaticBatching;
 
         public IMetaData MetaData { get => _metaDataName; set => _metaDataName = (MetaDataName)value; }
 
-        public CombineRendererBuilder(string parentName, int capacity = 0, int layer = 0)
+        public CombineRendererBuilder(string parentName, int capacity = 0, int layer = 0, bool applyStaticBatching = true)
         {
+            
             _materialCombineInstanceDictionary = new Dictionary<Material, List<CombineInstance>>(capacity);
             _materialCombineInstanceDictionary.TrimExcess();
 
-            _metaDataName = new MetaDataName(parentName);
             _layer = layer;
+            _applyStaticBatching = applyStaticBatching;
+
+            _metaDataName = new MetaDataName(parentName);
         }
 
         public IMetaDataNode<NodeType, MetaDataType> Create<NodeType, MetaDataType>(IMetaDataNode<NodeType, MetaDataType> parent) where MetaDataType : IMetaData
         {
             DataNodeBase combineRendererBuilder = (new GameObject(_metaDataName.Name)).AddComponent<DataNode>();
+            combineRendererBuilder.Header = ((DataNodeBase)parent).Header;
+            combineRendererBuilder.Parent = ((DataNodeBase)parent);
 
             if (_materialCombineInstanceDictionary != null) 
             {
                 foreach (KeyValuePair<Material, List<CombineInstance>> keyValuePair in _materialCombineInstanceDictionary)
                 {
                     DataNodeBase combineInstanceNode = (new GameObject(keyValuePair.Key.name)).AddComponent<DataNode>();
-                    combineInstanceNode.transform.parent = combineRendererBuilder.transform;
+                    combineInstanceNode.Header = combineRendererBuilder.Header;
+                    combineInstanceNode.Parent = combineRendererBuilder;
                     combineInstanceNode.gameObject.layer = _layer;
 
                     Mesh combinedMesh = new Mesh();
@@ -43,16 +51,12 @@ namespace DTTUnityCore.DataStructs
                 }
             }
 
+            if(_applyStaticBatching)
+            {
+                StaticBatchingUtility.Combine(combineRendererBuilder.gameObject);
+            }
+
             return (IMetaDataNode<NodeType, MetaDataType>)combineRendererBuilder;
-        }
-
-        public void Initialize(int capacity, string name = "Combine Renderers", int layer = 0)
-        {
-            _materialCombineInstanceDictionary = new Dictionary<Material, List<CombineInstance>>(capacity);
-            _materialCombineInstanceDictionary.TrimExcess();
-
-            _metaDataName = new MetaDataName(name);
-            _layer = layer;
         }
 
         public void AddCombineInstance(Material material, Mesh mesh, Matrix4x4 transform)